Converting PartialSumVec & its test to UME Simd backend: first prototype.
In order to investigate feasability of using vector operations with uint64_t ( i.e. usability of
its UInt64_v ), converted PartialSumVec & its test (TestVecPartialSum) to use UME Simd backend.

The goal is to create a first vectorised implementation of MIXMAX for a single stream on top of
VecCore - and the Partial Sum is a key building block.

- Revised top CMakeLists.txt making UME Simd (UMESIMD) the default backend.

- Revised PartialSumVec.h and TestVecPartialSum.cxx for use with UME Simd.

Status: compiles and runs.  Results are wrong.